[HUDI-6116] Optimize log block reading by removing seeks to check corrupted blocks.
The HoodieLogFileReader code currently does an eager isBlockCorrupted for which we do a seek and then a read of the footer. This invalidates internal buffers in opened file stream to the log file and makes a call to DataNode to start a new blockReader. Since the majority of the use case will not have corrupted blocks, the seek penalty is removed in favor of more aggressive checks for corrupted fields.